requirements.txt
python==3.10
PyPDF2==3.0.1
pdfplumber==0.8.0
nltk==3.8.1
scikit-learn==1.2.2
networkx==3.0
matplotlib==3.7.1
spacy==3.5.0
pyyaml==6.0
flask==2.2.3
python-dotenv==1.0.0
plotly==5.11.0
pypdf==3.10.0
beautifulsoup4==4.11.2
faiss-cpu==1.7.3
langchain==0.0.143
langchain-openai==0.0.4
langchain_community==0.0.5
rank-bm25==0.2.2
pymupdf==1.22.3
deepeval==0.1.1

# To download the spacy model after installation
spacy download en_core_web_sm
